Move logic to EmbroideryCanvas struct
