+++ src/api/mod.rs
@@ -1,4 +1,4 @@
 mod image;
-mod stitching_image;
 pub mod routes;
 pub mod colors;
+mod embroidery_canvas;
